requirements.txt
python-whois
time
sys
os
pathlib
coverage
